Moved test scan to the ekfs test
